Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cbd1f55e-20a0-4d78-a9f5-e41c17b4c8e7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5430355514a92881889ef61a048d3c0bbb72faee
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com