Explicitly use implementations where necessary for modules in the same apex
The logic that magically determines whether to use stubs or
implementations for modules in the same apex is being removed
to simplify the Soong logic and make the build system easier to
understand.  Explicitly use libdexfile#impl for modules in the
art apex that need the implementation APIs.

Bug: 372543712
Test: builds
Flag: EXEMPT refactor
Change-Id: I916d54fc558d6030bcb3b0586eb0f11a41a24c42