Revert "Reland^3 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/660750b3-bad1-49c7-a0d7-7be6489187c3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b3d9a08a98698a7ee431bffc0cc7327e6b51736
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com