Improve magnitudes recorded by Java Heap Profiling.
Reuse the same sampling logic as native heap profiling for consistency
and because it includes logic to scale up the magnitudes based on
sampling frequency.

Allocate TLAB just large enough for each object, to force the slow path
everywhere. The performance overheads of Java Heap Profiling are
dominated by the time to take a sample anyway.

Test: Manually profile an app with a specific allocation pattern and
verify magnitudes of reported allocations are much improved.

Bug: 370562214
Change-Id: Iba448509c107023f5d41dd1968f8bde0937bba51