Revert "Run RTP after GVN to remove more NullCheck instructions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed12654f-5cb5-4523-b06e-ef2b3777e835). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8d91b391c9548ed7abf588f6460bee4a0f0713e1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com