Introduce `RegType::Kind`.
Define the enumeration of all concrete `RegType` subclasses
and store that "kind" as a member `RegType`. Replace virtual
`Is##kind()` functions with a simple check of that member.
This allows the compiler to better optimize the code and
also avoid difficult to predict indirect jumps.

Rename some concrete `RegType` subclasses to align naming
with the `Is##kind()` functions. Rename `Is##kind()` for
`UnresolvedAndUninitialized*` to drop the `And`.

Remove redunant `IsLong()` and `IsDouble()` and use the
equivalent `IsLongLo()` and `IsDoubleLo()`, respetively.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 181943478
Change-Id: Ib075809b4244687cce1848c0d087e97e6f51cc98