Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b361100a-eeec-4777-87a1-29e450c8cdd3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8219ed6eb587cd0070250521535cf52058797c9d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com