Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1a3b0d07-464b-409a-a0c0-72f1db29c93a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I755051e2326f900cd6696f182e4e49671d6b0a5e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com