Revert "Poll for boot image mapping in a heap task."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/29858224-063f-4d2a-a62f-77cb14cd93ef). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I49044221eb79ee25508be290d2570271cc15082b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com