Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/754314fe-983a-4e56-ba23-f3bb74395ffb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie1a0c91158dc9c058c7c6b6b5dad4ec1746da57d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com