Revert "Poll for boot image mapping in a heap task."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/754314fe-983a-4e56-ba23-f3bb74395ffb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5d71ab921e3330bb87cba0c1f7746f144e58a33a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com