Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a77ae9f-1e5e-41d4-b74a-2b464f6c85d1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I99ed4a5ca79337152d3bd54084710969fed1ddd0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com