Revert "run-test: Create runner scripts as part of the build."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb706702-26ba-48d9-8370-4e04776ab31d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ica4a1c7086d48c60e2f6168e133a59561b036e7a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com