Revert "LUCI: Remove deprecated and hidden builders."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/697cf09f-db01-4f14-9723-7088d6989de2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 378141888
Change-Id: I5a297218dcc031de38f0dd4f3439cf65ed87ee70
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com