Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bc4a7ddd-0bf6-4b78-9bf5-641b6b729315). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6ae72bdb66aa950ddf1aea0bbce8d01266705ea4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com