Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f4523edf-77b8-465d-8742-2fbdd8684797). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6e55484fa9d5f96bc5a51e7d63ce02a46b42ba7c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com