Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2e8e84de-b546-4bb9-91ae-a707e074654c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I204a9bf55648bd11e7e06454132cdf1bc08bce1e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com