Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9f5c0b44-7f2b-45be-8861-b924008f5da0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie554e5e53fff8002b11c33f03acf5278e2c802d9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com