Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4bd89fd7-e77a-4b41-a781-97bd5654f543). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6eb227acf00b65a504cb255458054615bfad7cdf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com