Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fdc9c7ce-3fcd-4cf5-b3dc-3858eb4854aa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8fd22626659b1d5a76a9cc81a8fc54257b774c01
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com