Revert "run-test: Remove ART_HOST_TEST_DIR from env.py"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ee672b44-cd68-436a-a824-6294d82e6eb4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icbe73d4751cde57070b4827ea7dc9e9d3043bf79
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com