Revert "Increase black-dense threshold to 95%"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ee672b44-cd68-436a-a824-6294d82e6eb4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4616d17cefbfe3a2ad3f247707f16181cb9341bd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com