Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dc7d2ed6-891b-47af-b9bb-5839312e328e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I66099f901e84eb736d559dff1ddbba40b4bd8c34
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com