Revert "Replace test_for properties with explicit dependencies o..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ff16bc9a-7ed0-4ab4-9154-d8f255b94d74). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 378167290
Change-Id: I5456fb3a307c2d20e45e4b4590114d3e797b9681
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com