Revert "Consolidate the use of some constants."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c31d26d4-4599-43e8-a3fb-3b7f29f6db6d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 378457938
Change-Id: I9c47ec73489b178dc8d1637a23a9e8b2314c21b7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com