Clarify GetExclusiveOwner() semantics
It is inherently unreliable unless the mutex is held.
Contention reports may this err if the mutex has already been, or is
about to be, released.

Fixing this would require including exclusive_owner_ in the atomic
operation part of lock/unlock operations, which would require us to rely
on wider atomic operations on the fast path. So just document it better.

Bug: 377068860
Test: Treehugger
Change-Id: Id2be3dca573656e7aa10cf975f9a6c9275cd3541