Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dba9d711-ce76-427c-a22d-0ff70fb23ee8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I175cb060f687ad8a15c3034f346bf6a5696691ea
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com