Revert "Poll for boot image mapping in a heap task."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/264d099a-c01e-45ba-ac91-415a61decc7e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I42ba034b4976df8c0fe0ee70bba4cef4c5bb57a5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com