Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/264d099a-c01e-45ba-ac91-415a61decc7e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie2357fe1aae07dd19e03d727ef9da48067224317
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com