Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cc8025d1-611e-46f4-a15f-d6bbccf88c55). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I03ac1966627c85f302109b166fa7eba22687971b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com