Revert "Poll for boot image mapping in a heap task."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cc8025d1-611e-46f4-a15f-d6bbccf88c55). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I32d37b05a1ec8b56a89e4059ae8b8b5b2295b17e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com