Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d91a56ef-f79e-441a-a33d-e501e82680bf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7b85de68d352ea98a85bdedc9ed024e8ab7d1300
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com