Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1a8381f2-de53-49fb-9a06-3018022bee40). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I71d1e56b6a79a5cab11f6908fcd6684c38cfd6b4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com