Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ea8115f2-0a0a-4349-b31f-4f2bb7f89776). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I76f820c9410b154b31b995a95a9d94434ae2cc51
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com