Revert "Poll for boot image mapping in a heap task."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bfb83c47-7feb-4f25-b503-78ab5578432d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I734ec5a524fe740de226847a5926453bf0456ddf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com