Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bfb83c47-7feb-4f25-b503-78ab5578432d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I00c59e9801a2e81dc338a5a95fad837e78632297
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com