Revert "Don't block for long in UnloadNativeLibraries()"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6238d805-7a2a-4bea-809c-cb89f9524bab). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I484d7280915274a0b378e5ab8d94e3e6798459f8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com