Revert "Add a comment to explain the API considerations for the ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/601239d9-ae1d-4b80-99a0-6eab061f3291). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I98bc0a8130911183b57a0a3f98c08fa3f741134c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com