Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d94e9a8e-e9d0-45a9-9d95-0ac23052e24f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5eedece354e19281e77a8831c072bd9d93ad704d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com