Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/224413cd-6ed4-4217-b8bd-072a086d667f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie7d388c169c417c6d2864ee0a2bff2d82a9927a2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com