Revert "Poll for boot image mapping in a heap task."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/224413cd-6ed4-4217-b8bd-072a086d667f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia31483d5427e959ea821bccd5e7406e5452bcd77
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com