Revert "Add a comment to explain the API considerations for the ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1d4e145a-0133-4dff-a2a4-22a2128cf568). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9c2c7b8f84a25be14e812b80e41243945df60272
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com