Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b3b0d33b-f46d-494f-8aaa-a238c2ed3482). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2bc0bcbdfd1eb14400fe883373a550089c63320a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com