Revert "Poll for boot image mapping in a heap task."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/66cc9b59-3ecf-42d9-8c9b-21cfae803073). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I965d8f0fb6da5d889b90732fa0178ddb36df52fd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com