Revert "Poll for boot image mapping in a heap task."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c03b45c-a0d5-4fb7-8b10-5978c15d1b54). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b28a3786bf243a97627733dcd10bbe745afdc7a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com