Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c03b45c-a0d5-4fb7-8b10-5978c15d1b54). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22bf8f512194ee92eeb05b102d39295053909d13
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com