Revert "LUCI: Enable git superproject checkout for QEMU (non-repo)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ccacc30-a30b-44c6-899e-1f7765daf89d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib2c1115ab1636b8624954e3c4684cfd6cf2436c2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com