Revert "LUCI: Enable git superproject checkout for QEMU (non-repo)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de3258bb-b60a-464b-9ba4-2600752d366c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I918b96be806871de6caa4a3facbb3567fcde1484
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com