Revert "Export ART_TEST_ON_VM for tests."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fffe8b62-7cc5-4b61-a616-1b4ba2c2b8dd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7af1f5931da073c8401e24d65d0a7dee4fbd1614
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com