Revert "Replace test_for properties with explicit dependencies o..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ce650fcd-1dda-4b9c-9681-f8dd307a81a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4b3833cf5af29359464294a699327b1bf5dd13dd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com