Revert "Add a comment to explain the API considerations for the ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d597e49e-d8fe-429a-8f49-2ef10577e07f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I671131f2f6cb0e373439cd27543614d4d0c152b7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com