Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d0cfd4d5-19cf-4f51-a4ba-e5d04a78240b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id291991dcc1530049c0b59875be8c0ed3e4ca5ab
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com