Revert "Poll for boot image mapping in a heap task."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/608ae883-d398-4896-82e1-7f5f01990c62). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icfc9ba6beff4c47fc57e5f12a109b0cda08d51bc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com