Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/608ae883-d398-4896-82e1-7f5f01990c62). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6afdb8699a3612caff47b8481ecb072aa10ef9e0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com