Revert "Reduce the likelihood of the child process being stuck"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cc4c1624-6fc3-440e-9781-f2949af81c72). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I92a9fda68157856279b7d698b2be573c786dfdff
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com