Revert "Replace test_for properties with explicit dependencies o..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/95d066a3-f6ad-403b-a877-a83cbe4d6985). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I00a7e70f0f0c2981cb9abeb7fbd8f627ecf121a7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com