Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ae66463-5a4d-4c58-aa4d-f407d5ec866a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie268ddb9bb01d979bd4b9564324da8cd6d1429eb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com