Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d68cd72-f1fc-45db-a4de-25ce5d802156). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I141f70dedd96cf635eff33e6d90ea323c0bced08
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com