Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/feba9b05-24b4-4edd-a34a-9beba13b71bb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I77daf5f11e326fdce388cfb677b7e4509a3f7165
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com