Revert "LUCI: Enable git superproject checkout for QEMU (non-repo)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8e64428-9263-4121-aea2-9681b0355ea9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I47f0dcae30a3efc57193a6c0857ea94fe0c66064
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com