Revert "Poll for boot image mapping in a heap task."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ce1a51c-21d5-40a2-9e65-bb2109834863). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ief3fa1c63cd91839e884fa25eb1ded44062fcbc9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com