Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ce1a51c-21d5-40a2-9e65-bb2109834863). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7f6f954ec4bc9cc3871ec1e63bfe9b59d4732412
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com