Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd87a412-2ede-4ca7-b00b-0bf7f0aa31f5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icc17eff616cdb7bf577ca71a8946381bd2fd8a6e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com