Revert "Add implementation for Unsafe.compareAndExchangeLong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6e0c9a7-8501-450e-9d63-8a60e5c669ec). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I642feae23cf87cf34a435a3608a89499e4bc39ec
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com