Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4b3e9d65-438a-488e-a8d7-846bdc1cbc71). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If3ddf98c71b0b8fe66ff99fdaf8bdbd75b9c8cce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com