Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5e0363a1-2605-4110-9b47-bf6f0d09b08f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia5a5c3f4017c91aa4f8dabb6792ccd655a942223
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com