Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2593ed3f-fdb2-4ba1-99c1-808d81b73ac8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I726cb68cb430011bd4b93a4e0bd6f6f10f51e20e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com