Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dfdc25a6-c47e-4522-abac-6f81959c196d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I593a8f6f3f45e55de1f45f190cd1a87814cdd48f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com