Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f855ae6-4568-4ef8-b5c3-6bcda4dd0421). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id89f96c4ccfc3358312e226ceb3a5d2244c21ba5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com