Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/895294ba-2a1f-4d58-8fa4-ea64f1fa1f22). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib442e90cfdca99259378d98b9ac7b74a6fddef7b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com