Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f4c1e14-e1d3-4819-8df8-7a4b41fa947c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I532da3326fff284f280411f519a1f78e7c6e8744
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com