Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/778c7725-dd83-4021-a933-972f344849a6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I281fba8a43cd7cce40af91efb0f0982f042dea8a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com