Revert "Add a comment to explain the API considerations for the ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bee6966b-eea2-41f0-b589-ee7a0b53b18b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0bdcb84df34b5624d95e2e6e98f65179ab4bf863
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com