Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/453469c2-45c0-48e5-9a56-5619151cbc94). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iedd0c185881f7a71e75f580757ce8f8bb741da18
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com