Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1c808471-5257-4f9a-905c-460d314e2aed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If4a0eed41ebb9a266fd501260e152ba0ac2cefaf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com