Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9382c39a-96aa-4638-8bf8-1519ee7d4689). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I43ec6ca29b19261a571b29c2ed2d5f230a24bb62
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com