Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9d0599e5-b5ea-42f6-a037-fdc638172d96). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib601ae8ef96c76293e8bbf7537d59d721ee6b278
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com