Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/14601fd7-5613-41c6-9fb0-f52441bc18a4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6258c227349f71ed7b9a54ef2f8be5d9144f0e7e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com