Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/60a224ef-23ab-4e7e-b118-89dc4f24961f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie8157bc65bedd7e7d38b7b8000e121d0ea9560d4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com