Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5b2d3baf-17a4-4ba7-92a4-53155f5b969c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iabb198ae792d05629ad4cb3062c9eeb15365019b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com