Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f07396c7-c74f-4f68-8fd0-dee5cca92d00). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie7e56ccc11d7949f5a80143d8ab53f4c8f6aa41e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com