Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bfc81627-a4a5-40ab-b10e-df11d3348291). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I47c25c8f3a541de5cc1850b8a0f7324633c96f12
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com