Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f8e236d4-eb58-4492-9b37-05aa8bc6b8b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie991e74f39a774f803ce34096bd40aa4aba2a885
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com