Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/450a9b7f-78c9-405f-a7d4-8e35fed23371). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I451d59538cfc1b878a1e9a3aedc70e9d4e7afb94
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com