Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ac14e1d-9023-473e-938e-11d1cd38ca50). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8f13272518fb6edd1bbb203dc1293b9bd4ec30ac
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com