Revert "Allow the inliner to devirtualize intrinsics"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4acc8bb-79bd-4749-a69b-4f0b90cffd64). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic7465a679e5d06de1447f27748e957bdd7b6e0a0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com