Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/541c5b8b-abac-4922-8262-de4320cc6ad4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaf12b7bb43c55290b4a0d09ac22869d543fb6c9f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com