verifier: Stronger uninitialized `this` access checks.
Check uninitialized `this` field access early and make the
check more strict than before.

There are two cases that are now rejected and the run-test
064-field-access is updated with the corresponding tests.
First, we no longer allow iget on uninitialized `this`.
Second, we do not allow iput on uninitialized `this` to
specify a field by referencing a subclass even if the
field id actually resolves to the class being verified.

Note that doing this check early can also results in
reporting bad uninitialized `this` access instead of
another hard failure when both apply. It also suppresses
soft failures that would have been reported previously.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Change-Id: Ieb41115e6aaf618bc283f3aebf0739b606caa3b3