Revert "Avoid `strlen()` for `ClassLinker::FindClass()`..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a7bcc79b-b4fa-4380-b043-ba1e0a02fca0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9886efd7bf74f4362b38d5652eadc4aefb0c329b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com