Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a8c4315-c27c-4676-b919-97aa42884b57). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I057c6a0f58ec39f69f3b26a0390bfdac9d719312
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com