Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55315fdd-e2a4-4b64-9b3c-26b2304b0922). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2f0c6bd4dd494f443699c47c4696132dc1cdebc6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com