Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c392baef-fe8e-41a0-a7e9-67231d9392a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I65bd60a19811e712b5e7457649b21467c311991e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com