Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/73ea74cb-9b4b-49ea-b00f-302a24c5078a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icdad89922654a18f40722537ecb770396365229d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com