Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c8d73b1a-e5ab-4b29-9afc-aa570bab7ae5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I91578ec925f8c604d9dedb2848283beb01ffabd6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com