Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d3d87522-67ff-48f7-91a5-3cc89597a946). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5ffed67c38fd780ca104f11ad7c7a36e639931f6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com