Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dcd7b319-ef44-408a-a86f-71d4d6759017). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5c1ccca692ec2d1bc2700017cb4c014d081dce4b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com