Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55a122bd-7325-4e21-b041-cf01cd763f0c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If3e18973caae5eb7be861ca4a9e73143e065751f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com