Revert "Allow the inliner to devirtualize intrinsics"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2d3e0f23-21b2-4b90-b96e-8134ede799bf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5a5e8f1e00b16744c2bec275aceb57f926b35b57
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com