Revert "verifier: Refactor type verification."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2d3e0f23-21b2-4b90-b96e-8134ede799bf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I882ed83b5bb624722296019d11a477a6f0e952a2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com