Revert "Allow the inliner to devirtualize intrinsics"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c4ba1bea-8d6a-487f-989e-e0ee3ecb04b5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id4f2ad9a82c6cc8b6c8a3ebd55e8869c0b82a6c0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com