Revert "verifier: Refactor type verification."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c4ba1bea-8d6a-487f-989e-e0ee3ecb04b5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I38b1a226da3f9ce7250ad3735088250f5b864f94
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com