Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55ae32bc-05ba-4220-a0b3-f2ed8bd702d9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie07080a2951d4f06e23b3628e398b7386f9a1d36
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com