Revert "verifier: Refactor type verification."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ccadc5d2-9d3f-4e83-b8f6-0b6246482c5b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib480239e2c560aab5875db375cad6351da052023
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com