Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a6ad6ad-da0e-4dbe-a4a8-a83f1df692bd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I06c2429652a409cd5bcc5b072468d7eae0e17d12
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com