Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0fcb5137-ac86-41a6-853f-6504efada5ec). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieb6552cc4a1cf59db859c2bc51023b2209d36feb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com