Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/40500a46-445a-44b0-908a-84e2e9518c6d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I420d77a0e4920ee45d29eac33845a37ea14fef30
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com