Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/712e0cef-f09e-4a99-8903-db122caeecf2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifac2d333d1568083e2452fc78c75920b92ce0d41
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com