Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a84cf076-9613-4f11-8276-b47421ab78e5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I90f4946ef32d296a42efb1838f475ea5d73d8b5c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com