Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9459df75-9f66-41d6-ba97-df2a03185a38). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I88b49210d22de031d9b7f2c29ca43bf36fdc6613
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com