Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a78c898-acea-4c01-bf0d-b6947055933a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic3e02ed35ce323472b0ece5749422a36aecaf75d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com